chore(backend): address PR #49 review — test rename + AAA + scope env vars
Deep-review findings addressed in one pass against the Unit 1 persistence
foundation:

- Rename StartupSmokeTests.cs → StartupSmokeIntegrationTests.cs (file + class)
  so integration tests match backend/CLAUDE.md's {Class}IntegrationTests.cs
  convention; the class is [Trait(Category, Integration)] + uses
  WebApplicationFactory<Program> + Testcontainers so the integration suffix is
  unambiguous.
- Add // Arrange / // Act / // Assert markers + actualXxx locals to every test
  in the class (seven of nine were missing them); two that already had markers
  stay as-is for narrative consistency.
- Add SutHost_DevelopmentMigrationService_Is_Registered_As_HostedService_In_Development
  — catches a regression that silently drops the AddHostedService<
  DevelopmentMigrationService>() call in Program.cs. The fixture pre-migrates
  via a throwaway DbContext so StartAsync finds no pending migrations at
  runtime, which is why a registration-level assertion (not a behavioral one)
  is the practical coverage under the current test topology.
- RunCoachAppFactory.InitializeAsync now delegates to the existing
  CreateDbContext() helper instead of duplicating the DbContextOptionsBuilder
  pattern — single source of truth for the test-side options wiring.
- Replace historyCount.Should().BeOfType<long>().And.Subject.As<long>().Should().
  BeGreaterThan(0, …) with .Should().BeOfType<long>().Which.Should().
  BeGreaterThan(0, …) — the idiomatic FluentAssertions form with no .Which
  re-entry.
- Retire SmokeTests.cs — StartupSmokeIntegrationTests.Health_Endpoint_Returns_Ok_Json
  subsumes its single opt-in GetHealth_ReturnsOk assertion with stronger
  checks (content type + JSON body shape), eliminating the AssemblyFixture
  env-var side-effect coupling that opt-in test had.

CodeRabbit nits addressed in the same pass:

- Scope RunCoachAppFactory's process-wide env-var override with save/restore:
  InitializeAsync captures the prior ConnectionStrings__runcoach /
  OTEL_EXPORTER_OTLP_ENDPOINT values before overwrite; DisposeAsync restores
  them. Keeps the env-var path R-055 confirmed works (ConfigureAppConfiguration
  / UseSetting did not reliably beat the JSON providers on this stack) while
  eliminating the cross-contamination concern CodeRabbit flagged.
- DpKeysContext.DataProtectionKeys is now a => Set<DataProtectionKey>()
  expression-bodied accessor rather than a public {get; set;} = null! field;
  matches the RunCoachDbContext pattern and removes null-forgiveness.

Verified locally: 582 passing / 0 failing / 0 skipped in ~2s (was 581 / 0 / 1
before review round).
